Moved all handlers to handlers/samples directory
