+++ config/functions/samples.yml
@@ -10,13 +10,66 @@ sampleCache:
         private: false
         cors: ${self:custom.http.cors}
 
-errorHandling:
+sampleErrorHandling:
   handler: ${self:custom.path.app}/handlers/samples/errorHandling.handler
   description: Error Handling sample test
   reservedConcurrency: 1
   events:
     - http:
         path: samples/error-handling
+        method: get
+        private: false
+        cors: ${self:custom.http.cors}
+
+sampleQueue:
+  handler: ${self:custom.path.app}/handlers/samples/queue.handler
+  description: Queue sample test
+  memorySize: 192
+  reservedConcurrency: 1
+  events:
+    - http:
+        path: samples/queue
+        method: get
+        private: true
+        cors: ${self:custom.http.cors}
+
+sampleQueueProcessor:
+  handler: ${self:custom.path.app}/handlers/samples/queueProcessor.handler
+  description: Process sample queued test
+  memorySize: 128
+  reservedConcurrency: 3
+  events:
+    - sqs: arn:aws:sqs:${self:provider.region}:${env:AWS_ACCOUNT_ID}:${self:provider.stackName}-sampleQueue
+
+sampleScheduledEvent:
+  handler: ${self:custom.path.app}/handlers/samples/scheduledEvent.handler
+  description: Ping event triggered from Scheduled Cloudwatch Event
+  events:
+    - schedule:
+        name: ${self:provider.stackName}-sampleScheduledEvent
+        description: Trigger sampleScheduledEvent lambda function every 5 minute
+        rate: rate(5 minutes)
+        enabled: false
+
+sampleObjectStore:
+  handler: ${self:custom.path.app}/handlers/samples/objectStore.handler
+  description: S3 Sample
+  reservedConcurrency: 1
+  events:
+    - http:
+        path: samples/object-store
+        method: get
+        private: false
+        cors: ${self:custom.http.cors}
+
+sampleJwt:
+  handler: ${self:custom.path.app}/handlers/samples/jwt.handler
+  description: Sample jwt test
+  timeout: 5
+  reservedConcurrency: 1
+  events:
+    - http:
+        path: samples/jwt
         method: get
         private: false
         cors: ${self:custom.http.cors}
@@ -45,55 +98,3 @@ errorHandling:
 #         method: get
 #         private: false
 #         cors: ${self:custom.http.cors}
-
-# objectStore:
-#   handler: ${self:custom.path.app}/handlers/samples/objectStore.handler
-#   description: S3 Sample
-#   reservedConcurrency: 1
-#   events:
-#     - http:
-#         path: samples/objectStore
-#         method: get
-#         private: false
-#         cors: ${self:custom.http.cors}
-
-# queue:
-#   handler: ${self:custom.path.app}/handlers/samples/queue.handler
-#   description: SQS Sample
-#   reservedConcurrency: 1
-#   events:
-#     - http:
-#         path: samples/queue
-#         method: get
-#         private: false
-#         cors: ${self:custom.http.cors}
-
-# pingQueue:
-#   handler: ${self:custom.path.app}/handlers/pingQueue.handler
-#   description: Ping queued test
-#   memorySize: 192
-#   reservedConcurrency: 1
-#   events:
-#     - http:
-#         path: ping/queue
-#         method: get
-#         private: true
-#         cors: ${self:custom.http.cors}
-
-# pingQueueProcessor:
-#   handler: ${self:custom.path.app}/handlers/pingQueueProcessor.handler
-#   description: Process queued ping
-#   memorySize: 128
-#   reservedConcurrency: 3
-#   events:
-#     - sqs: arn:aws:sqs:${self:provider.region}:${env:AWS_ACCOUNT_ID}:${self:provider.stackName}-pingQueue
-
-# pingScheduledEvent:
-#   handler: ${self:custom.path.app}/handlers/pingScheduledEvent.handler
-#   description: Ping event triggered from Scheduled Cloudwatch Event
-#   events:
-#     - schedule:
-#         name: ${self:provider.stackName}-pingScheduledEvent
-#         description: Trigger pingScheduledEvent lambda function every 5 minute
-#         rate: rate(5 minutes)
-#         enabled: false
