+++ src/handlers/samples/queueProcessor.js
@@ -1,7 +1,14 @@
+import middy from 'middy';
+import normalizeSQSMessageMiddleware from 'Middlewares/normalizeSQSMessageMiddleware';
+import { connectSentry } from 'Utils/sentry';
 import ErrorException from 'Exceptions/ErrorException';
 import { PING_QUEUE_PROCESSOR_SAMPLE_ERROR } from 'Constants/errorCodes';
 
-const pingQueueProcessor = collection => {
+connectSentry();
+
+const originalHandler = event => {
+  const { collection } = event;
+
   if (collection[0].data === undefined || collection[0].data === null)
     return collection;
 
@@ -14,4 +21,7 @@ const pingQueueProcessor = collection => {
   );
 };
 
-export default pingQueueProcessor;
+// eslint-disable-next-line import/prefer-default-export
+export const handler = middy(originalHandler);
+
+handler.use(normalizeSQSMessageMiddleware());
