+++ src/handlers/samples/scheduledEvent.js
@@ -6,6 +6,6 @@ connectSentry();
 
 // eslint-disable-next-line import/prefer-default-export
 export const handler = event => {
-  logger.info('received pingScheduled Event', { event });
+  logger.info('Received Scheduled Event', { event });
   return ping();
 };
